backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
openpyxl==3.1.5
google-genai==1.5.0
python-dotenv==1.0.1
Pillow>=11.1.0
